feat(rt): add initial support for streaming HTTP request/response bodies (#19)
This change introduces a `ByteStream` type that any blob shape with a streaming trait can target during codegen. Similar in nature to `HttpBody` but meant for more general purpose use in a generated SDK.

* Add streaming support for the Ktor engine
* Always represent response payloads as a stream and let the pipeline decide whether or not to pull it all into memory (e.g. for deserialization) or consume it as a stream.
* Related changes: POC example(s) are now a direct subproject of smithy-kotlin for now

This allows client-runtime changes to be immediately consumed and a better debugging experience. This will make the feedback loop tighter while we figure out the client-runtime. The POC project will get removed from the main project at a later time.

Co-authored-by: Aaron J Todd <todaaron@amazon.com>