feat(rt): generic deserializer interface and JSON Implementation (#21)
This commit adds a generic deserializer interface (the analogous flip side to #18) with support for the JSON format.

## Background
Deserialization is quite a bit more involved than serialization since we don't know what the payload contains, the order fields will come in, dealing with unknown fields, etc.

There are two common approaches to deserialization and dealing with this complexity:

1. Raw token iteration
The idea is you iterate over a stream of tokens and store whatever state you need to along the way until you have either exhausted the token stream or built the structure you wanted. Tokens are usually very low level like "found start of array" or "found number". Compilers do this as part of the lexing and AST construction process for example. 

2. Visitor
The other common way to deal with deserialization is to implement the visitor pattern. You override methods such as `visitString(...)` or `visitInt(...)` etc. The biggest difference here is you as the consumer don't drive the deserialization process, instead the deserializer calls you with what it found. 

Both methods work but usually aren't the primary way users want to interact with deserialization since both methods are overly verbose and easy to misuse. Where you usually find direct interaction with these approaches is processing very large XML or JSON documents that can't be pulled into memory directly and you only care about sub-parts of the document.

In the JVM world (and many other languages) users interact with deserialization through reflection. Properties of classes are marked with annotations and the deserialization framework uses reflection capabilities to construct an object from a given payload (using perhaps either (1) or (2) in the process). Instead of driving a token iterator or visitor a user can simply say "give me a Foo object". Both GSON and Jackson work this way (in addition to exposing a raw token stream).

## Implementation Details

The approach taken here is very similar to how kotlinx.serialization works. Instead of iterating raw (protocol specific) tokens, you provide the deserializer with metadata about what you expect to find (ex `SdkFieldDescriptor`) and then it tells you what it found at each iteration.

This approach has the nice property of being rather friendly for codegen purposes. The interface for deserialization is format agnostic (meaning you don't force this complexity into the code generator to deal with different token streams) yet high level enough that you are mostly matching the expected class property types to deserializer methods (`fieldx` is an int -> `deserializeInt()`). In other words it turns raw token iteration into field iteration.

The v2 Java SDK is actually similar in nature but drastically different in execution. It too takes the approach of providing the field metadata (via `SdkField`) to the deserializer but instead of generating a deserializer implementation for the specific type it stuffs more information into the metadata and lets the deserializer build the object (via setters).

The v1 Java SDK and Android SDK seem to implement more of a raw token iteration.