fix: bad rebase/merge. serde interfaces changed
Commits: 0cbb164:c3b4ff3 were based on the original serde
interfaces and I failed to fix the interface changes that happened in f5c36ed.
These commits will never work as is until this fixme commit. Normally
we'd redo this but it's not worth the effort. It's early days and these
in-between commits will never be tagged or revisited as is.