refactor: rename dslBuilder() -> builder() (#253)
Minor usability improvement that gives favor of the simple builder name to the Kotlin DSL style builder instead of the Java fluent builder API. This will hopefully make the interface a little more clear at first glance to new users and keep them from accidentally using the fluent API because they miss the DSL builder. I left the name of the type as DslBuilder as I think it actually aids in understanding.