fix: remove maxTime from StandardRetryStrategy
Removes a maximum per/operation timeout from the StandardRetryStrategy.
There is no sensible default for this timeout since an operation could
involve a large request stream. It also subjected operations to timeout
in highly concurrent environments with lots of coroutines if the
coroutine was starved and just didn't get a chance to run.

fixes: awslabs/aws-sdk-kotlin/issues/572