refactor(rt): make ExecutionContext implement CoroutineScope
Most background work is related to HTTP engine but there are cases
where clients need a way to launch background work related to request
processing (e.g. event streams). This change makes ExecutionContext
implement CoroutineScope for background work to be launched into.
This scope is cancelled and joined when an SdkHttpOperation is done
executing.