fix(rt): OkHttp request body retry failing due to Job completed (#832)
When OkHttp config has retry on connect failures enabled it will attempt
to re-use the request body in some scenarios. The coroutine Job was
setup to only be used once though and was completed when the original
request fails. This causes the (internal okhttp) retry to fail with an obscure
JobCompleted exception.